Fixes potential overflow in memory allocation
Casts width to size_t when allocating memory to prevent
integer overflow and ensure correct allocation size,
improving robustness when handling large images.

+++ library/sgl-core/resize/generic_nearest_neighbor.c
@@ -29,7 +29,7 @@ sgl_result_t sgl_generic_resize_nearest(uint8_t *dst, int32_t d_width, int32_t d
     }
 
     if (errcnt == 0) {
-        x_lookup = (int32_t *)malloc(sizeof(int32_t) * d_width);
+        x_lookup = (int32_t *)malloc(sizeof(int32_t) * (size_t)d_width);
         if (x_lookup != NULL) {
             /* create 'column' lookup table */
             for (col = 0; col < d_width; ++col) {
